+++ spec/models/dataset/value_spec.rb
@@ -81,4 +81,45 @@ RSpec.describe Dataset::Value, type: :model do
       end
     end
   end
+
+  describe 'value=' do
+    context 'when format is a select' do
+      before do
+        value.key = FactoryBot.build(:dataset_key, :select)
+      end
+
+      it 'accept an array' do
+        value.value = %w[a b]
+        expect(value.value).to eq('a, b')
+      end
+
+      it 'ignores empty array values' do
+        value.value = %w[a] + [nil] + %w[c]
+        expect(value.value).to eq('a, c')
+      end
+
+      it 'accepts as string' do
+        value.value = 'a'
+        expect(value.value).to eq('a')
+      end
+    end
+  end
+
+  describe 'mapped_value' do
+    context 'when format is a boolean' do
+      before do
+        value.key = FactoryBot.build(:dataset_key, :boolean)
+      end
+
+      it 'returns No for value 0' do
+        value.value = 0
+        expect(value.mapped_value).to eq('No')
+      end
+
+      it 'returns Yes for value 1' do
+        value.value = 1
+        expect(value.mapped_value).to eq('Yes')
+      end
+    end
+  end
 end